Refactor junit RunnableSuite.
