+++ modules/core-cats/shared/src/main/scala/weaver/BaseIOSuite.scala
@@ -2,6 +2,6 @@ package weaver
 
 import cats.effect.IO
 
-trait BaseIOSuite extends RunnableSuite[IO] with BaseCatsSuite {
+trait BaseIOSuite extends EffectSuite[IO] with BaseCatsSuite {
   protected def effectCompat: UnsafeRun[IO] = CatsUnsafeRun
 }
