+++ modules/core/js-native/src/main/scala/weaver/SharedResourceRunnableSuite.scala
@@ -1,0 +1,6 @@
+package weaver
+
+/**
+ * Stub implementation. The JS and Native SharedResourceSuite is not runnable.
+ */
+private[weaver] trait SharedResourceRunnableSuite[F[_]] extends EffectSuite[F]
