--- modules/core/js/src/main/scala/org/junit/runner/RunWith.scala
@@ -1,9 +1,0 @@
-package org.junit.runner
-
-import org.typelevel.scalaccompat.annotation.unused
-
-/**
- * Stub used for cross-compilation
- */
-class RunWith[T](@unused cls: Class[T])
-    extends scala.annotation.StaticAnnotation
