--- modules/core/js/src/main/scala/weaver/junit/WeaverRunner.scala
@@ -1,6 +1,0 @@
-package weaver.junit
-
-/**
- * Stub used for cross-compilation
- */
-class WeaverRunner()
