+++ modules/discipline/js-native/src/main/scala/weaver/discipline/DisciplineFRunnableSuite.scala
@@ -1,0 +1,5 @@
+package weaver
+package discipline
+
+/** Stub implementation. The JS and Native DisciplineFSuite is not runnable. */
+private[discipline] trait DisciplineFRunnableSuite[F[_]] extends EffectSuite[F]
